refactor tests to reuse auth state and common methods/items
