+++ strr-host-pm-web/tests/e2e/smoke-tests/Scenario-2.test.ts
