+++ strr-host-pm-web/tests/e2e/test-utils/constants.ts
@@ -1,0 +1,2 @@
+import { LoginSource } from '../enums/login-source'
+export const loginMethods = [LoginSource.BCSC, LoginSource.BCEID]
